Stabilize Pi skill command coverage
