+++ plugins/provider-pi/src/bridge/bridge.ts
@@ -27,7 +27,6 @@ import {
   providerInstallationStatusParamsSchema,
   providerMaintenanceParamsSchema,
   isStandaloneBuiltinCompactCommand,
-  mimeTypeFromExtension,
   modelListParamsSchema,
   runBridgeRequest,
   skillsConfigureParamsSchema,
@@ -78,6 +77,7 @@ import {
   resolvePiBridgeSessionDir,
   resolvePiSessionFilePath,
 } from "./session-paths.js";
+import { extractPiPromptInput } from "./turn-input.js";
 
 const piCommandSchema = z.discriminatedUnion("method", [
   z.object({
@@ -1106,7 +1106,7 @@ async function handleTurnStart(
     sendResult(id, { threadId: params.threadId });
     return;
   }
-  const { text, images } = extractInput(params.input);
+  const { text, images } = extractPiPromptInput(params.input);
   if (!text && images.length === 0) {
     sendError(id, BRIDGE_JSON_RPC_ERRORS.INVALID_PARAMS, "Missing input text");
     return;
@@ -1133,7 +1133,7 @@ async function handleTurnSteer(
     sendError(id, -32000, "No active pi session");
     return;
   }
-  const { text, images } = extractInput(params.input);
+  const { text, images } = extractPiPromptInput(params.input);
   if (!text && images.length === 0) {
     sendError(id, BRIDGE_JSON_RPC_ERRORS.INVALID_PARAMS, "Missing input text");
     return;
@@ -1197,79 +1197,6 @@ async function handleThreadDiscard(
     },
   );
   return { ok: true };
-}
-
-interface ExtractedInput {
-  text?: string;
-  images: ImageContent[];
-}
-
-interface SelectedPiSkill {
-  chunkIndex: number;
-  end: number;
-  name: string;
-  start: number;
-}
-
-function extractInput(input: TurnStartParams["input"]): ExtractedInput {
-  const chunks: string[] = [];
-  const images: ImageContent[] = [];
-  const skills: SelectedPiSkill[] = [];
-  for (const item of input) {
-    if (!item || typeof item !== "object") continue;
-    const typed = item as {
-      type?: string;
-      text?: string;
-      path?: string;
-      mimeType?: string;
-    };
-    if (typed.type === "text" && typeof typed.text === "string") {
-      const chunkIndex = chunks.push(typed.text) - 1;
-      for (const mention of item.type === "text" ? item.mentions : []) {
-        const resource = mention.resource;
-        if (
-          resource.kind === "command" &&
-          resource.source === "skill" &&
-          resource.trigger === "/" &&
-          mention.start >= 0 &&
-          mention.start < mention.end &&
-          mention.end <= typed.text.length &&
-          typed.text.slice(mention.start, mention.end) ===
-            `${resource.trigger}${resource.name}`
-        ) {
-          skills.push({
-            chunkIndex,
-            end: mention.end,
-            name: resource.name,
-            start: mention.start,
-          });
-        }
-      }
-    } else if (typed.type === "localImage" && typeof typed.path === "string") {
-      try {
-        const data = readFileSync(typed.path).toString("base64");
-        const mimeType = typed.mimeType ?? mimeTypeFromExtension(typed.path);
-        images.push({ type: "image", data, mimeType });
-      } catch {}
-    } else if (typed.type === "localFile" && typeof typed.path === "string") {
-      chunks.push(`[Attached file: ${typed.path}]`);
-    }
-  }
-  const [skill] = skills;
-  if (skills.length === 1 && skill) {
-    const chunk = chunks[skill.chunkIndex];
-    if (chunk !== undefined) {
-      chunks[skill.chunkIndex] =
-        `${chunk.slice(0, skill.start)}${chunk.slice(skill.end)}`;
-      const argumentsText = chunks.join("\n");
-      const separator = argumentsText.startsWith(" ") ? "" : " ";
-      return {
-        text: `/skill:${skill.name}${argumentsText ? `${separator}${argumentsText}` : ""}`,
-        images,
-      };
-    }
-  }
-  return { text: chunks.length > 0 ? chunks.join("\n") : undefined, images };
 }
 
 function handleParsedMessage(parsed: unknown): void {
